+++ src/components/WalletModal/WalletModal.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from "react";
+import React, { useCallback, useState, useMemo } from "react";
 
 import BigNumber from "bignumber.js";
 import { useWallet } from "use-wallet";
@@ -12,14 +12,14 @@ import Grid from "@mui/material/Grid";
 
 import FancyValue from "components/FancyValue";
 import useBalances from "hooks/useBalances";
-import { addInkToWallet, addGransToWallet, addwUbqToWallet, addEschToWallet } from "metamask.js";
+import { addTokenToWallet } from "metamask.js";
 
-import { Logos } from "farms/AvailableFarms";
+import { Tokens } from "farms/AvailableFarms";
 
 const WalletModal: React.FC<ModalProps> = ({ isOpen, onDismiss }) => {
   const [, setWalletModalIsOpen] = useState(false);
   const { reset } = useWallet();
-  const { INKBalance, UBQBalance, GRANSBalance, ESCHBalance } = useBalances();
+  const { tokenBalances } = useBalances();
 
   const handleSignOut = useCallback(() => {
     localStorage.removeItem("account");
@@ -31,17 +31,26 @@ const WalletModal: React.FC<ModalProps> = ({ isOpen, onDismiss }) => {
     }
   }, [reset, onDismiss]);
 
+  const WalletTokens = Tokens.map((token, index) => {
+    const balance = tokenBalances === undefined ? undefined : tokenBalances[token.address];
+
+    return (
+      <WalletToken
+        tokenName={token.symbol}
+        tokenIcon={token.publicLogo}
+        tokenOnClick={() => addTokenToWallet(token.symbol, token.address, token.publicLogo)}
+        tokenBalance={balance}
+        key={index.toString()}
+      />
+    );
+  });
+
   return (
     <Modal isOpen={isOpen}>
       <ModalTitle text="My Wallet" />
       <ModalContent>
         <Box>
-          <Grid>
-            <WalletToken tokenName={"wUBQ"} tokenOnClick={addwUbqToWallet} tokenBalance={UBQBalance} />
-            <WalletToken tokenName={"INK"} tokenOnClick={addInkToWallet} tokenBalance={INKBalance} />
-            <WalletToken tokenName={"GRANS"} tokenOnClick={addGransToWallet} tokenBalance={GRANSBalance} />
-            <WalletToken tokenName={"ESCH"} tokenOnClick={addEschToWallet} tokenBalance={ESCHBalance} />
-          </Grid>
+          <Grid>{WalletTokens}</Grid>
         </Box>
       </ModalContent>
       <Separator />
@@ -59,12 +68,23 @@ const WalletModal: React.FC<ModalProps> = ({ isOpen, onDismiss }) => {
 
 interface WalletTokenProps {
   tokenName: string;
+  tokenIcon: string;
   tokenOnClick?: () => void;
   tokenBalance: BigNumber | undefined;
 }
 
-const WalletToken: React.FC<WalletTokenProps> = ({ tokenName, tokenOnClick, tokenBalance }) => {
+const WalletToken: React.FC<WalletTokenProps> = ({ tokenName, tokenIcon, tokenOnClick, tokenBalance }) => {
   const getDisplayBalance = useCallback((value?: BigNumber) => {
+    if (value !== undefined) {
+      if (value.isLessThan(0.0001) && value.isGreaterThan(0)) {
+        return value.toExponential();
+      }
+
+      if (value.isEqualTo(0)) {
+        return "0";
+      }
+    }
+
     if (value) {
       return numeral(value).format("0.0000a");
     } else {
@@ -72,18 +92,28 @@ const WalletToken: React.FC<WalletTokenProps> = ({ tokenName, tokenOnClick, toke
     }
   }, []);
 
-  const AddTokenButton = function () {
+  const AddTokenButton = () => {
     if (tokenOnClick !== undefined) {
       return <Button onClick={tokenOnClick}>Add {tokenName} to Wallet</Button>;
     }
     return <></>;
   };
 
+  const TokenImg = useMemo(() => {
+    return (
+      <img
+        src={`/tokens/${tokenIcon}`}
+        alt={`${tokenName} token logo`}
+        style={{ marginRight: "10px", height: 64, alignSelf: "center", background: "white", borderRadius: 110 }}
+      />
+    );
+  }, [tokenIcon, tokenName]);
+
   return (
     <Box sx={{ display: "flex" }}>
       <Grid container>
         <Grid item xs={7}>
-          <FancyValue icon={getIcon(tokenName)} label={tokenName + " balance"} value={getDisplayBalance(tokenBalance)} />
+          <FancyValue icon={TokenImg} label={tokenName + " balance"} value={getDisplayBalance(tokenBalance)} />
         </Grid>
         <Grid item xs={5}>
           <AddTokenButton />
@@ -92,33 +122,5 @@ const WalletToken: React.FC<WalletTokenProps> = ({ tokenName, tokenOnClick, toke
     </Box>
   );
 };
-
-function getIcon(logo: string) {
-  let icon;
-  switch (logo) {
-    case "INK":
-      icon = Logos.ink;
-      break;
-    case "wUBQ":
-      icon = Logos.ubq;
-      break;
-    case "GRANS":
-      icon = Logos.grans;
-      break;
-    case "ESCH":
-      icon = Logos.esch;
-      break;
-    default:
-      console.warn("getIcon() logo does not exist", logo);
-  }
-
-  return (
-    <img
-      src={icon}
-      alt={`${logo} token logo`}
-      style={{ marginRight: "10px", height: 64, alignSelf: "center", background: "white", borderRadius: 110 }}
-    />
-  );
-}
 
 export default WalletModal;